scripts/requirements.txt
# 核心（向量检索 MCP Server + 索引构建）
FlagEmbedding>=1.2.0
qdrant-client>=1.9.0
mcp[cli]>=1.0.0
python-frontmatter>=1.1.0
torch>=2.0.0
numpy>=1.24.0
